+++ plugins/spamassassin.js
@@ -18,6 +18,10 @@ exports.load_spamassassin_ini = function () {
             '+check.private_ip',
             '+check.local_ip',
             '+check.relay',
+
+            '-defer.error',
+            '-defer.connect_timeout',
+            '-defer.scan_timeout',
         ],
     }, () => {
         plugin.load_spamassassin_ini();
@@ -274,6 +278,8 @@ exports.get_spamd_headers = function (conn, username) {
 
 exports.get_spamd_socket = function (next, conn, headers) {
     const plugin = this;
+    const txn = conn.transaction;
+
     // TODO: support multiple spamd backends
 
     const socket = new sock.Socket();
@@ -282,7 +288,7 @@ exports.get_spamd_socket = function (next, conn, headers) {
 
     socket.on('connect', function () {
         // Abort if the transaction is gone
-        if (!conn.transaction) {
+        if (!txn) {
             plugin.logwarn(conn, 'Transaction gone, cancelling SPAMD connection');
             socket.end();
             return;
@@ -296,20 +302,22 @@ exports.get_spamd_socket = function (next, conn, headers) {
     });
 
     socket.on('error', err => {
-        conn.logerror(plugin, `connection failed: ${err}`);
-        // TODO: optionally DENYSOFT
-        // TODO: add a transaction note
+        socket.destroy();
+        if (txn) txn.results.add(plugin, {err: `socket error: ${err.message}` });
+        if (plugin.cfg.defer.error) return next(DENYSOFT, 'spamd scan error');
         return next();
     });
 
     socket.on('timeout', function () {
+        socket.destroy();
         if (!this.is_connected) {
-            conn.logerror(plugin, 'spamd connection timed out');
+            if (txn) txn.results.add(plugin, {err: `socket connect timeout` });
+            if (plugin.cfg.defer.connect_timeout) return next(DENYSOFT, 'spamd connect timeout');
         }
         else {
-            conn.logerror(plugin, 'timeout waiting for results');
+            if (txn) txn.results.add(plugin, {err: `timeout waiting for results` });
+            if (plugin.cfg.defer.scan_timeout) return next(DENYSOFT, 'spamd scan timeout');
         }
-        socket.end();
         return next();
     });